remove compiler warnings for real to in conversions
git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@46021 6189c7af-7406-0410-a2ce-f22f51069c23

Old repository SHA1: eb7b1877e6efbdff34def110a0f075423a4840f8

+++ rosetta_source/src/protocols/swa/rna/StepWiseRNA_Dinucleotide_Sampler_Util.cc
@@ -396,12 +396,12 @@ namespace rna {
 		int const euler_angle_bin_min=-180/euler_angle_bin_size; //Should be -180/euler_angle_bin_size
 		int const euler_angle_bin_max=180/euler_angle_bin_size-1;  //Should be 180/euler_angle_bin_size-1
 
-		int const euler_z_bin_min=-1/euler_z_bin_size;
-		int const euler_z_bin_max=(1/euler_z_bin_size)-1;
+		int const euler_z_bin_min=-int(1/euler_z_bin_size);
+		int const euler_z_bin_max= int(1/euler_z_bin_size)-1;
 
 		Real const max_distance=12;
-		int const centroid_bin_min=-max_distance/static_cast<Real>(centroid_bin_size);
-		int const centroid_bin_max=(max_distance/static_cast<Real>(centroid_bin_size))-1;
+		int const centroid_bin_min = -int(max_distance/static_cast<Real>(centroid_bin_size));
+		int const centroid_bin_max =  int((max_distance/static_cast<Real>(centroid_bin_size))-1);
 
 		std::cout << "euler_angle_bin min= " << euler_angle_bin_min << " max " << euler_angle_bin_max << std::endl;
 		std::cout << "euler_z_bin_min min= " << euler_z_bin_min << " max " << euler_z_bin_max << std::endl;
